faculties/pyproject.toml
[tool.poetry]
name = "faculties"
version = "0.1.0"
description = ""
authors = ["Yartsev Vladimir <v.rainovsky@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
strawberry-graphql = {extras = ["fastapi"], version = "^0.133.6"}
uvicorn = "^0.18.3"
sqlalchemy = {extras = ["aiosqlite"], version = "^1.4.41"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
